Extract `printBlock`
