+++ src/language-js/printer-estree.js
@@ -123,6 +123,7 @@ const {
 const { printBinaryishExpression } = require("./print/binaryish");
 const { printStatementSequence } = require("./print/statement");
 const { printMemberExpression } = require("./print/member");
+const { printBlock } = require("./print/block");
 const { printComment } = require("./print/comment");
 
 function genericPrint(path, options, printPath, args) {
@@ -514,65 +515,8 @@ function printPathNoParens(path, options, print, args) {
       return "import";
     case "TSModuleBlock":
     case "BlockStatement":
-    case "StaticBlock": {
-      const naked = path.call((bodyPath) => {
-        return printStatementSequence(bodyPath, options, print);
-      }, "body");
-
-      if (n.type === "StaticBlock") {
-        parts.push("static ");
-      }
-
-      const hasContent = n.body.some((node) => node.type !== "EmptyStatement");
-      const hasDirectives = n.directives && n.directives.length > 0;
-
-      const parent = path.getParentNode();
-      const parentParent = path.getParentNode(1);
-      if (
-        !hasContent &&
-        !hasDirectives &&
-        !hasDanglingComments(n) &&
-        (parent.type === "ArrowFunctionExpression" ||
-          parent.type === "FunctionExpression" ||
-          parent.type === "FunctionDeclaration" ||
-          parent.type === "ObjectMethod" ||
-          parent.type === "ClassMethod" ||
-          parent.type === "ClassPrivateMethod" ||
-          parent.type === "ForStatement" ||
-          parent.type === "WhileStatement" ||
-          parent.type === "DoWhileStatement" ||
-          parent.type === "DoExpression" ||
-          (parent.type === "CatchClause" && !parentParent.finalizer) ||
-          parent.type === "TSModuleDeclaration" ||
-          parent.type === "TSDeclareFunction" ||
-          n.type === "StaticBlock")
-      ) {
-        return concat([...parts, "{}"]);
-      }
-
-      parts.push("{");
-
-      // Babel 6
-      if (hasDirectives) {
-        path.each((childPath) => {
-          parts.push(indent(concat([hardline, print(childPath), semi])));
-          if (
-            isNextLineEmpty(options.originalText, childPath.getValue(), locEnd)
-          ) {
-            parts.push(hardline);
-          }
-        }, "directives");
-      }
-
-      if (hasContent) {
-        parts.push(indent(concat([hardline, naked])));
-      }
-
-      parts.push(comments.printDanglingComments(path, options));
-      parts.push(hardline, "}");
-
-      return concat(parts);
-    }
+    case "StaticBlock":
+      return printBlock(path, options, print);
     case "ThrowStatement":
     case "ReturnStatement":
       return concat([
